+++ packages/datasources/src/index.ts
@@ -3,3 +3,4 @@ export * from "./drizzle.js";
 export * from "./prisma.js";
 export * from "./supabase.js";
 export type { default as ITransactor } from "./transactor.interface.js";
+export * from "./typeorm.js";