+++ packages/datasources/src/typeorm.ts
@@ -1,0 +1,2 @@
+export { default as TypeOrmDatasource, type TypeOrmContext } from "./typeorm.datasource.js";
+export { default as TypeOrmTransactor } from "./typeorm.transactor.js";