+++ packages/filters/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./drizzle/index.js";
 export * from "./prisma/index.js";
 export * from "./supabase/index.js";
+export * from "./typeorm/index.js";
 export type { FieldOperators, Operator, Order, OrderDirection, QueryFilters, Scalar, ScalarKeys, ScalarSelect, Select, Unwrap, Where } from "./types.js";